requirements.txt
playwright==1.28.*
requests==2.*
percy-playwright
